More details on "why bodyguards fail to spawn" (and why 841e62a may not be the best fix)
There are other lines of code in the scripts that call
> (assign, "$g_mt_mode"
and assign values of 1.

In particular:

- abm_training
- ctm_melee
- vba_normal

The vba_normal is used when you enter a village
infested with bandits (a common situation for players
that would trigger this bug).

It looks like very few of those scripts set the global
variable back to value zero after the situation.

This commit tries changing the tcm_disguised constant
to a more unique value like 99.  That removes any
conflict with the other possible values in that global
variable.