EXP: Adjust the price adjustment as you buy goods The scripts "game_event_buy_item" and "game_event_sell_item" are what adjusts the price up/down as the player purchases things from the merchant. This only kicks in for trade goods and not armor/weapons/horses.
The constant "price_adjustment" does not seem to be used anywhere
in the scripts and definitely not in the two game_event scripts.  It might
be used internally by the engine.

The math for this is:

1. Get the current price multiplier for the item at the party.  This
gives a value around 1000 (100% of the static item value) but
could range from 700-1300 (min is 100, max is 10000).

2. Always add/subtract a static value of 10-30 to the multiplier.
This has the effect that each additional item purchased is 1-3%
more expensive. This commit reduces it from 2-3% change to
a 1-2% change.

3. For items above 100 (now 150) item_value, subtract that from the
item_value, then divide by 15 (was 8).  Add/subtract this new value to
the multiplier.

OLD BUY:
1200 + 20 + (380 - 100)/8 = 1255 (4.5% increase)
1000 + 20 + (380 - 100)/8 = 1055 (5.5% increase)
800 + 20 + (380 - 100)/8 = 855 (6.9% increase)

NEW BUY:
1200 + 10 + (380 - 150)/15 = 1225 (2.1% increase)
1000 + 10 + (380 - 150)/15 = 1025 (2.5% increase)
800 + 10 + (380 - 150)/15 = 825 (3.1% increase)

The end result of this change is a slower change in the price
as you buy out the merchant's inventory.  Higher priced goods
are still more volatile, but not as much.