Require higher rep with lords/towns to purchase If DPLMC_GOLD_CHANGES_HIGH is active:
In order to purchase businesses, you must now raise your reputation
with both the town and the town's liege.  The amount varies based
on the liege's type and the party's trade skill.

If DPLMC_GOLD_CHANGES_MEDIUM is active, you still need +1
with the town and now you also need +2 with the liege.