rework the "charity" (boost low-prosp villages) There's still some more ideas that I have, but this now takes away 300g from either the lord of the village or the liege of the faction before boosting prosperity by 1-5 points.
Acts as another gold sink.  Only removes money from
the lord if they have 5000+ denars (or 10k for the kingdom
liege).  May want to reduce the boost if the NPC can't pay?

Need to figure out a way for the player to take advantage
of the system.  Maybe a separate charity account that
the player pays into?  Or they can use one of their NPCs
(chancellor?) to enable charity for the villages.