rebalance bandit spawns and lair spawns Old code was guaranteed to add one bandit party spawn per bandit party type every 36h.  New code has a 50% chance every 13h.  Which is a bit faster.
Old code had a static spawn limit for the bandit parties,
new code uses a limit of 6 parties if the bandit-lair is
destroyed and increases the limit by +12 if the lair exists.
This does allow the player to have a significant impact on
the number of bandit parties running around and interfering
with farmers going to/from market.

Old code had a 100% chance to attempt a respawn of the
lair every 36h.  There's only a ~30-50% chance of success
on each attempt due to restrictions on terrain, x/y/z coord
values, etc.  New code runs every 13h, but with a 20%
chance.  That means rougly 2.7 day between attempts,
which is down for 1.5 days between attempts.

Deserters are no longer a static limit of 15 on the map,
but are now 8 + (# of wars x 8).  This may be too much
and would need to scale back to a x5 or x6 based on
the number of wars.

Looter spawn limit is no longer static, but is also based
on the number of wars.  Looters tend to spawn after
walled fiefs change hands, so the looter party count
naturally ends up above the limit most of the time.