rework village/town goods merchant restocking Number of slots that get stocked with goods is now primarily driven by prosperity, with a penalty for being besieged/infested/looted/deserted/etc. status.
Goods will be stocked if the base production of the
local trade network is >0, but at a very low probability.

Goods that have positive net production, get a higher
probability.  Expensive goods get a lower probability.

In general, villages that are normal-status will have 12-29
slots filled with goods.  Towns will be 34-83 slots.