return the average price-index for a good When calculating market-village production, it also calculates the average price-index for the good. In the case of villages, only villages where the farmer made it to market in the last 144 hours are counted.
The price-index for the good is averaged with the
average_price_factor to tilt it towards the average_price_factor
before mixing it into the other village values.  Thus, the
price-index returned will lie halfway between the actual
village averages and the average_price_factor value.

The theory is that if a good has 600 index in the village,
this will increase to 800 (halfway to 1000) by the time
it makes it to the market town.