refactor do_party_center_trade - only tariff positive differences in the price-index - sqrt() the price-index diff between origin/dest - take the average of each profitable trade-good - randomize the price-index diffs - add a base of 100 denars - divide final result by party-type (6 for farmers, 2 for caravans)
This will give more even results for tariffs between towns
where one town has a very different average price-factor
value than another town.  It should even out tariffs a bit
so that (# of caravans) is the driving factor for tariffs and
not solely which town the caravan came from.

It still only does a price-index adjustment on some (1/3) of
the trade goods and not all trade goods.  Between that and
the change to how :percentage_change is calculated, it
allows towns to slowly move into agreement but not be
forced into immediate agreement on prices.