add more trade caravan routes between towns Some towns had under 10 routes, now all towns should have at least 10 routes and the central towns have 14-15 trade routes defined.
The routes should mostly make sense and allow a caravan
from town A to skip past town B to service town C.

Caravans still prefer their own faction 50% of the time and
consider distance 25% of the time.  So just because a
caravan could travel on Jelkala-Bariyye, does not mean
it will.