Slow down the pace of NPC courtship Lords like to visit their NPC love interest periodically and are limited by (hours_since_last_courtship), which is 168 hours. But the rate of increase for gain/loss with the lady was very fast, gaining as much as 20-25 points in a single visit. Since the mark for becoming betrothed is +20 (or -20 for either rejecting/forcing), this rate of increase needs slowed.
1. The 168 hour cooldown is still in place.  The script
"troop_does_business_in_center" handles this.  Search for
slot_party_leader_last_courted or the variable named
":hours_since_last_courtship".

2. For each time the lord attempts to do business in the center
and chooses to court the love interest, only 5% of the attempts
will succeed.  This is kind of brute-force, but definitely slows
down the rate at which courtship_event_troop_court_lady is
being called.  It's not clear whether the lord will give up and
leave the center after failure, or if they are locked in place
until they get a success.

3. The score from "lady_evaluate_troop_as_suitor" script now
uses larger divisors (was 5..15) to decrease how fast the two
NPCs move towards bethrothal.  The evaluation script seems
to return values between -120 and +120, but usually in the
-100 to +100 range.  The lower end of the divisor is now 10,
which means romantic chemistry below ~20 is not going to
be viable.

4. After the divisor is applied to the chemistry score, you're left
with a value that is approximately -12 to +12, but usually in
the -4 to +4 range.  This value is then fuzzed by picking a random
value between -4 and +2 and adding it.  This negative bias knocks
the borderline cases slowly into negative territory.

5. Lastly, the maximum number of relations points that can be gained
or lost is clamped to -3 to +3.  For high chemistry romances, they will
be constantly bumping against the +3 limit, which would be ~7 weeks
from zero to betrothal.  Slower courtships might take 20-25 weeks or
longer depending on random fuzzing.

If even slower courtship is desired, the -20 / +20 constants in the
courtship_event script may need to be adjusted.  There may also
need to be a way to break off courtship if the value never leaves
the -5 to +5 range for a while.  This is unlikely as lords/ladies will
often have 2-3 suitors each competing against each other.