Grain prod is double-impacted by prosp For the production of itm_grain, take the center's prosperity and adjust up/down by up to 50%, with prosp=50 being the "no impact" midpoint.
Also cut the probability that grain will appear in the village's
elder's inventory.  This inventory is checked when the elder
decides if they want to ask the player to deliver grain.  It was
proving difficult to find an elder with zero grain.

(There is a trick where the player buys out all the grain from
the elder, then asks the elder for a task.  But the grain quest
has a second restriction on the prosperity of the village, so
this trick only works on low-prosperity villages.)