Cleanup and adjust process_hero_ai Also start tweaking npc_decision_checklist_party_ai
process_hero_ai -- When a hostile lord is searching for
a nearby village to attack, they now flip a coin to process
the list of villages in forwards or backwards ordering. The
first village that is within 15 map units will be attacked.

npc_decision_checklist_party_ai

1) Lords that have a score to settle with an enemy will
randomize their relation with the compared-to lord by
+/- 10 relation points.  Only certain lord types will choose
this action.  But the player could still be at the top of
their hate-list, even with the randomization.

2) Lords with <1500 denars will raid for money (was 500).

3) Lords who are raiding for money are less fussy about
where they choose to raid.  So instead of taking the
village prosperity and subtracting distance, the two
values are randomized by +/- 30% before doing the
math.  It might even be worth increasing the fuzziness
to +/- 35% to +/- 50%.

4) Lords that are looking for the wealthiest lands use
a similar calculation, but only fuzz the prosperity values
by +/-20% (and distance by +/- 30%).

It may be better to randomize prosperity by points
instead of percentage.  That would require a val_clamp
on the prosperity value.