Player loses less money when village is looted Right now, the entire treasury is up for grabs when the player's villages are looted.  Which stops making sense as soon as the player has multiple fiefs (villages / towns / castles).  It's reasonable to assume that the towns and castles would hold more of the treasury, with less of it stored at the village.
Or, if the player only has villages, that they would spread
their risk around.

This calculates the number of shares (points) for fiefs that the
player controls.  Villages=1, Castles=2, Towns=4.  The treasury
is divided by the total number of shares (points) and if a center
is sacked, the player only loses one share.

Ex: (2) towns, (1) castle, (3) villages = 13 shares

In any case where one of the centers is razed/captured, the
player is only at risk of losing up to 1/13 of the treasury.