The profit score (cur_town_score) is now randomized a bit by -3 to +3 after calculating a value in the 7..17 range.
What was happening in-game was many of the caravans
on the map making a beeline for any recently besieged
town (due to the price increases at the town during the
siege).

The distance calculation has also been fuzzed.  It
divides the map distance by 50, which gives a value
of 0..6, boosted to 3..9, which is then fuzzed by a random
value in -2 to +2.